feat: update router types export
