+++ packages/plugin-router/src/runtime/Router.tsx
@@ -9,7 +9,8 @@ import {
 
   RouteComponentProps
 } from 'react-router-dom';
-import { RoutesProps, RouterProps, IRouteWrapper, IDynamicImportComponent, RouteItemProps, IRenderRouteProps } from '../types';
+import { RoutesProps, RouterProps, IRouteWrapper, IDynamicImportComponent, RouteItemProps } from '../types/router';
+import { IRouterConfig } from '../types';
 
 function wrapperRoute(component, routerWrappers) {
   return (routerWrappers || []).reduce((acc, curr) => {
@@ -37,7 +38,7 @@ function getRouteComponent(component, routerWrappers?: IRouteWrapper[]) {
 function parseRoutes(routes: RouteItemProps[]) {
   return routes.map((route) => {
     const { children, component, routeWrappers, ...others } = route;
-    const parsedRoute: IRenderRouteProps = { ...others };
+    const parsedRoute: IRouterConfig = { ...others };
     if (component) {
       parsedRoute.component = getRouteComponent(component, children ? [] : routeWrappers);
     }
@@ -56,7 +57,7 @@ export function IceRouter(props: RouterProps) {
   const parsedRoutes = parseRoutes(routes);
   return (
     <RouterComponent {...others}>
-      <Routes routes={parsedRoutes} fallback={fallback}/>
+      <Routes routes={parsedRoutes} fallback={fallback} />
     </RouterComponent>
   );
 }
