+++ packages/plugin-router/src/types/collector.ts
@@ -1,0 +1,17 @@
+
+export interface ICollectItem {
+  routePath: string;
+  component: string;
+  filePath: string;
+  isLayoutLike: boolean;
+  exact?: string;
+  routePathAmend?: string;
+  children?: ICollectItem[];
+}
+
+export interface IIgore {
+  pattern: RegExp;
+  attributes?: string;
+}
+export type IgnoreType = string | IIgore;
+export type IgnoreOptions = IgnoreType | IgnoreType[];
