+++ packages/plugin-router/src/types/index.ts
@@ -4,16 +4,20 @@ import {
 } from 'react-router-dom';
 import { History } from 'history';
 
-export interface IDynamicImportComponent {
+interface IModifyRoutes {
+  (modifyFn: IModifyFn): void;
+}
+
+interface IRouteWrapper {
+  (props: any): React.ComponentType<any>;
+}
+
+interface IDynamicImportComponent {
   __LAZY__: boolean;
   dynamicImport: () => Promise<{ default: React.ComponentType<any> }>;
 }
 
-export interface IRouteWrapper {
-  (props: any): React.ComponentType<any>;
-}
-
-export interface RouteItemProps extends DefaultRouteProps {
+interface RouteItemProps extends DefaultRouteProps {
   children?: RouteItemProps[];
   // disable string[]
   path?: string;
@@ -25,40 +29,8 @@ export interface RouteItemProps extends DefaultRouteProps {
   routeWrappers?: IRouteWrapper[];
 };
 
-export interface IRenderRouteProps extends DefaultRouteProps {
-  children?: IRenderRouteProps[];
-  // disable string[]
-  path?: string;
-  // for rediect ability
-  redirect?: string;
-
-  component?: React.ComponentType<RouteComponentProps<any>> | React.ComponentType<any>;
-}
-
-export interface RouterProps {
-  // custom props
-  routes: RouteItemProps[];
-  type?: 'hash' | 'browser' | 'memory' | 'static';
-  // common props for BrowserRouter&HashRouter&MemoryRouter
-  basename?: string;
-  getUserConfirmation?: ((message: string, callback: (ok: boolean) => void) => void);
-  forceRefresh?: boolean;
-  // for BrowserRouter
-  keyLength?: number;
-  // for HashRouter
-  hashType?: 'slash' | 'noslash' | 'hashbang';
-  // for MemoryRouter
-  initialEntries?: string[];
-  initialIndex?: number;
-  fallback?: React.ReactNode;
-};
-
 interface IModifyFn {
   (routes: RouteItemProps[]): RouteItemProps[];
-}
-
-export interface IModifyRoutes {
-  (modifyFn: IModifyFn): void;
 }
 
 export interface IAppRouterProps {
@@ -70,32 +42,12 @@ export interface IAppRouterProps {
   history?: History;
 }
 
-export interface RoutesProps {
-  routes: IRenderRouteProps[];
-  fallback?: React.ReactNode;
-};
+export interface IRouterConfig extends DefaultRouteProps {
+  children?: IRouterConfig[];
+  // disable string[]
+  path?: string;
+  // for rediect ability
+  redirect?: string;
 
-export interface IRouterOptions {
-  caseSensitive?: boolean;
-  ignoreRoutes?: IgnoreOptions;
-  ignorePaths?: IgnoreOptions;
-  configPath?: string;
-  lazy?: boolean;
+  component?: React.ComponentType<RouteComponentProps<any>> | React.ComponentType<any>;
 }
-
-export interface ICollectItem {
-  routePath: string;
-  component: string;
-  filePath: string;
-  isLayoutLike: boolean;
-  exact?: string;
-  routePathAmend?: string;
-  children?: ICollectItem[];
-}
-
-export interface IIgore {
-  pattern: RegExp;
-  attributes?: string;
-}
-export type IgnoreType = string | IIgore;
-export type IgnoreOptions = IgnoreType | IgnoreType[];
